Moved remainig services into new service folder and removed the auth guard service.
