+++ src/app/_services/bookkeeping-service/bookkeeping.service.ts
@@ -18,7 +18,7 @@
 
 import { Injectable } from '@angular/core';
 import { IncomeStatementItem } from 'src/app/_models/IncomeStatementItem';
-import { ApiService } from '../../_services/api-service/api.service';
+import { ApiService } from '../api-service/api.service';
 import { BehaviorSubject, Observable, map, reduce } from 'rxjs';
 import { PaginationLoadData } from 'src/app/_models/PaginationLoadData';
