+++ src/app/_services/portfolio-service/portfolio.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { ApiService } from '../../_services/api-service/api.service';
+import { ApiService } from '../api-service/api.service';
 import { Asset } from 'src/app/_models/asset';
 
 @Injectable({
