+++ src/app/app-routing.module.ts
@@ -23,7 +23,6 @@ import { BookkeepingCreateComponent } from './pages/bookkeeping/bookkeeping-crea
 /* --------------------------------
     ACCESS GUARDS
 ---------------------------------*/
-import { authenticationGuard } from './services/authorization-services/auth-guard-service/auth-guard.service';
 
 const routes: Routes = [
     {
