+++ src/app/components/modal/bookkeeping-delete-modal-content/bookkeeping-delete-modal-content.component.ts
@@ -19,7 +19,7 @@ import { Component, Input } from '@angular/core';
 import { Router } from '@angular/router';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 import { ApiService } from 'src/app/_services/api-service/api.service';
-import { BookkeepingService } from 'src/app/services/bookkeeping/bookkeeping.service';
+import { BookkeepingService } from 'src/app/_services/bookkeeping-service/bookkeeping.service';
 
 @Component({
     selector: 'app-delete-modal-content',
