+++ src/app/pages/dashboard/dashboard.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Observable, Subscription } from 'rxjs';
 import { IncomeStatementItem } from 'src/app/_models/IncomeStatementItem';
-import { BookkeepingService } from 'src/app/services/bookkeeping/bookkeeping.service';
-import { ChartService } from 'src/app/services/chart/chart.service';
+import { BookkeepingService } from 'src/app/_services/bookkeeping-service/bookkeeping.service';
+import { ChartService } from 'src/app/_services/chart-service/chart.service';
 import { Chart } from 'chart.js';
 
 @Component({
