--- src/app/services/authorization-services/auth-guard-service/auth-guard.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { AuthGuardService } from './auth-guard.service';
-
-describe('AuthGuardService', () => {
-    let service: AuthGuardService;
-
-    beforeEach(() => {
-        TestBed.configureTestingModule({});
-        service = TestBed.inject(AuthGuardService);
-    });
-
-    it('should be created', () => {
-        expect(service).toBeTruthy();
-    });
-});
